+++ src/components/TargetResourceBrowser.js
@@ -3,8 +3,8 @@ import { keyframes } from "emotion";
 import { jsx } from "@emotion/core";
 import { useSelector, useDispatch } from "react-redux";
 import { actionCreators } from "../redux/actionCreators";
-import ResourceButton from "./widgets/ResourceButton";
-import TargetResourcesHeader from "./widgets/TargetResourcesHeader";
+import ResourceButton from "./widgets/buttons/ResourceButton";
+import TargetResourcesHeader from "./widgets/headers/TargetResourcesHeader";
 import textStyles from "../styles/text";
 import TargetSearch from "./TargetSearch";
 import Spinner from "./widgets/Spinner";
