+++ src/components/widgets/headers/MediumHeader.js
@@ -3,7 +3,7 @@ import { keyframes } from 'emotion';
 import { jsx } from '@emotion/core';
 import PropTypes from 'prop-types';
 
-import textStyles from '../../styles/text';
+import textStyles from '../../../styles/text';
 
 const fadeIn = keyframes`
   0% {
